Revert "Change type of maxBufferSize and remove include."
This reverts commit 0e13c5669719e250b49f425d1e1306f824c4823d.

+++ Includes/libARNetworkAL/ARNETWORKAL_Manager.h
@@ -8,6 +8,7 @@
 #ifndef _ARNETWORKAL_MANAGER_H_
 #define _ARNETWORKAL_MANAGER_H_
 
+#include <stdlib.h>
 #include <libARNetworkAL/ARNETWORKAL_Error.h>
 #include <libARNetworkAL/ARNETWORKAL_Frame.h>
 
@@ -140,7 +141,7 @@ struct ARNETWORKAL_Manager_t
     void *senderObject; /**< Internal reference, do not use */
     void *receiverObject; /**< Internal reference, do not use */
     int maxIds; /**< Maximum supported buffer ID for ARNetwork */
-    unsigned int maxBufferSize; /**< Maximum size allowed for a buffer, header included. */
+    size_t maxBufferSize; /**< Maximum size allowed for a buffer, header included. */
 };
 
 /**
